Remove resolveSuperTypes() and clean up after
The `Codebase.resolveSuperTypes()` method was needed when the
`ApiAnalyzer.propagateHiddenRemovedAndDocOnly()` computed the various
API variant selector properties like `hidden`, `docOnly`, `removed` so
that it would compute those properties for all classes in the class
hierarchy. However, now that those properties are computed dynamically
as needed `resolveSuperTypes()` is no longer needed.

This change removes the method and then does the following clean up:
* Remove `PsiBasedCodebase.finishInitialization()` and
  `ApiFile.postProcess()` as they are now empty.
* Remove `iterateAllClasses(...)`, `allClasses` and `addClass(...)` as
  they were only use by `resolveSuperTypes()` and so are no longer
  needed.

The `TextTypeParserCacheTest` needed updating because the total number
of classes in the `Codebase` is reduced because super types are no
longer being resolved.

Bug: 352480646
Test: ./gradlew
Change-Id: I0143c9a4bb125bab9149cbe42f0b40a720b0c772