Correct visibility of test suite tests
Previously, quite a lot of test suite tests used the wrong visibility
on classes which were meant to be public. That made them unrealistic
as if they were used to generate an API they would be ignored.

This change corrects the visibility in preparation for a follow-up
change that will ignore any class that is inaccessible unless it is a
super class for an accessible. That is because inaccessible classes
cannot contribute to an API directly and can only contribute through
providing methods and/or fields that are inherited into an
accessible class.

Bug: 325341742
Test: ./gradlew
Change-Id: I7e8862ba552aad7f87e78fa96d62f2d8bbd14219