Support origin correctly in TurbineCodebaseInitialiser
Previously, `TurbineCodebaseInitialiser` would assume that all classes
that were provided as source were provided on the command line and all
other classes were from the class path. It would not differentiate
between classes that were provided on the command line and classes that
were found on the source path. This change corrects that.

Bug: 354699754
Test: ./gradlew
Change-Id: I203c89ef75baba6c281caea2fa1fa7a0e69696da