TurbineCodebaseInitialiser: Scan source path
Previously, Turbine would only process the files that had been
explicitly specified on the command line. That differs from Psi which
will look on the source path for any additional files that it needs.
That discrepancy is one of the issues that prevents Turbine from being
used downstream in the Android build because it relies on that Psi
behavior. It does not have much impact on the tests as most tests will
pass all their inputs on the command line.

It causes issues downstream as Turbine fails because its binder cannot
resolve references to classes that are on the source path but not
explicitly specified. In that case the binder throws an exception.

This change makes Turbine scan the source path for all the `.java`
files it can find and then includes them in the classes it binds. That
should fix some of the unresolved class issues in the Android build.
However, it may not fix them all as there are no guarantees that the
`.java` files it finds are a self-consistent set, something which Psi
does not care about but Turbine does. Longer term it may be necessary
to make changes in the Android build to ensure that the `.java` files
provided to Metalava are self-consistent.

Bug: 354699754
Test: ./gradlew
Change-Id: I11b19153b325cccd3be115c986c935550878a24a