Remove duplicate NonNull annotations in AnnotationsMergerTest
`libcore.util.NonNull` is not repeatable so specifying it twice is not
strictly correct. Psi does not care about it but Turbine does. At the
moment this test passes on Turbine because Turbine does not see the
definition of the `libcore.util.NonNull` annotation because it is only
provided on the source path and not in the explicit list of source
files to process. As a result Turbine does not know if it is repeatable
so it does not throw an error

However, a future change will get Turbine to look for source files on
the source path and process them at which point it will be given the
definitions of `libcore.util.NonNull` and start reporting repeated
instances as an error. This change avoids that by removing one of the
instances.

Bug: 325341742
Test: ./gradlew
Change-Id: I6be9146ce244850aee94a7ace80fcf6ff0cbdaec