Use consistent order for stub methods
Previously, stub methods were written out in so-called "source order".
The primary reason for this was because the method order in the source
did affect the order of methods in the documentation but that is no
longer true. The "source order" was ill-defined as it depended on lots
of factors, i.e.
1. The order of methods within a single class.
2. The order of methods across classes (when methods were inherited
   from one class to another).
3. The order in which classes were created affected the order of
   methods across classes.
4. The order in which Metalava copied methods from one class to
   another.

Changing any of those caused unnecessary churn in the stubs, breaking
tests and requiring the build to do extra work. This change sorts them
consistently by the signature and fixes the tests.

Bug: 365745549
Test: ./gradlew
Change-Id: I6ff6eb87d2fe88c753fc7e62c2786e41ae8f98b9