Move JvmOverloads workaround for constructors to Psi model
Previously, the workaround for the Psi issue
https://youtrack.jetbrains.com/issue/KT-57537 was to emulate the
behavior of `@JvmOverloads` by creating overloads of the constructors,
one for each parameter with a default value. That was added to
`SignatureWriter` but that meant that it would not work with snapshots
as it relied on having a `PsiConstructorItem` when generating the
signature which is not true when using snapshots. It was also the wrong
place because a problem with the Psi model should be fixed in the Psi
model, and not leak through the model API.

This change moves the emulation of `@JvmOverloads` from
`SignatureWriter` to the Psi model where it belongs. As that is done
before any snapshot is taken that ensures that it will work properly
when using snapshots. That allowed the `shouldExpandOverloads()` to be
removed from the `CallableItem` and moved to be a private method in
`PsiCodebaseAssembler`.

Bug: 367623741
Test: ./gradlew
(cherry picked from https://android-review.googlesource.com/q/commit:e7edec3f7664cfc2c8a5bf3975081e9c4b9cb08a)
Merged-In: I73053d79ee1a8c12f1967d2349df6a09fffcaacc
Change-Id: I73053d79ee1a8c12f1967d2349df6a09fffcaacc