Remove CodebaseSnapshotTaker.typeItemFactory/Stack
Instead of trying to keep track of the current
`SnapshotTypeItemFactory` in the stack, which requires preserving class
nesting which prevents taking a snapshot of a partial `Codebase` this
creates a new `SnapshotTypeItemFactory` from the containing class. That
will be slightly more expensive but if it becomes a performance issue
then they can easily be cached.

Bug: 368021325
Test: ./gradlew
Change-Id: I418e46e24daaad1e5d5e27701968bb1051c4422c