Do not emit snapshot classes added when resolving references
Previously, the `emit` property of snapshot classes was determined by
their `ClassOrigin`, i.e. set to false if it was not `COMMAND_LINE`.
However, it is possible for classes with `ClassOrigin.COMMAND_LINE` to
be in the snapshot but not be allowed in the output. That occurs when
a type def annotation is private so the definition is not emitted but
the `typeDefMode` is `REFERENCE` so a reference to the annotation is
written out so it must be in the snapshot.

This change supports that by setting `emit = false` for any class that
is referenced from the snapshot but not defined in the snapshot.
Without this change following changes will break some tests.

Bug: 357833584
Test: ./gradlew
Change-Id: I9e2558f80d46a598855b72f3e95da0cced198896