Add SignatureToJDiff tests for adding new interface
Previous refactorings to how `ClassItem`s are registered has broken the
`SignatureToJDiffCommand`. This change adds tests to cover that
behavior for outermost and nested interfaces.

The lack of a test for something as simple and basic as adding a whole
new class is a big hole but it seems as though very few tests, if any,
were added when the functionality was first written and were only added
when some issue was detected.

A follow-up change will fix the issue.

Bug: 368634012
Test: ./gradlew
Change-Id: I45b3f6cad13eaeda1ea1c28c581cabc8da04885a