Fix handling of new class in SignatureToJDiffCommand
A previous refactoring of the class registration code (specifically
https://r.android.com/3184721 which moved the registration into the
`DefaultClassItem` initializer in order to ensure consistency) broke
the handling of new classes in `SignatureToJDiffCommand` when
generating a delta between two different APIs. The problem was that
while the diff-ing code detected that there was a new class and tried
to add it, it was not registered correctly which meant that it was
ignored when writing out the JDiff file.

This change duplicates some of the registration code from
`DefaultClassItem` initializer to fix the issue and corrects the newly
added broken tests that showed the problem.

A separate bug has been created to try and clean up the code as it
does not create a valid `Codebase`.

Bug: 368634012
Test: ./gradlew
Change-Id: Id2f46edd4de05695fba32161827f6d92651f78cb