Make AnnotationItem.targets consistent across models
Previously, the models supplied a `classFinder` lambda that was used to
retrieve a `ClassItem` on which a `Predicate<Item>` could be applied to
see which, if any `AnnotationTarget`s the annotation should be included
in. That `classFinder` was `resolveClass(String)` on Psi but
`findClass(String)` on Turbine which caused inconsistent behavior. This
change makes them consistent by using `resolveClass(String)` which
makes `classFinder` and the Psi specific overrides of `targets`
unnecessary so they are removed.

Bug: 370698104
Test: ./gradlew
Change-Id: Ie14fbc2bde0c2059c70a6ad04ef05568cc08c5c4