Add test to show getImports(...) methods ignore unresolved classes
The `getImports(Predicate<Item>)` method implementations use
`findClass(...)` to find an imported class to see whether it can be
referenced. If they cannot find the class then they will ignore the
import. That means that the behavior of the `getImports(...)` methods
is affected by whether a class has been resolved prior to that method
being called. That makes the behavior quite fragile as classes are only
resolved when needed and so performance optimizations that avoid
resolving a class that was previously resolved can change the behavior.

This change adds a test to detect that problem. The test is structured
very carefully to ensure that one of the imported classes is not
resolved. To ensure consistency the existing imports test was also
restructured in a similar way.

A follow-up change will fix this by switching to use `resolveClass`.

Bug: 371955471
Test: ./gradlew
Change-Id: I579aa3f2b6a73af3ab4136fafb57ae2288565443