Use DefaultAnnotationManager as default in testsuite
Previously, the test suite used `noOpAnnotationManager` as the default
`AnnotationManager`. This change switches it to use the
`DefaultAnnotationManager` as that is more realistic. While most tests
are not affected by the switch a number of tests are broken due to the
annotation name normalization that is performed. Those tests have been
modifier to explicitly use `noOpAnnotationManager`. That approach was
taken rather than change the test to use the normalized names as the
names are important to the tests.

Bug: 370045964
Test: ./gradlew
Change-Id: I5dcdb9d176847d41691f3e1333b94cfd96a31094