Snapshot AnnotationItem.info
Previously, the `AnnotationItem.snapshot(...)` method would not take a
copy of `AnnotationItem.info` which meant that it might not have been
initialized. If it was not initialized then the `AnnotationInfo` would
be created from the snapshot `AnnotationItem`. Unfortunately, the
`AnnotationInfo` properties depend on information, e.g.
`ApiVariantSelectors`, which are not available in the snapshot so if
they are computed from the snapshot they could end up with the wrong
values.

This change makes `AnnotationItem.snapshot(...)` ensure that the
`AnnotationItem.info` property is initialized so the corresponding
`AnnotationInfo` will be created from the original `AnnotationItem`.

Bug: 368021325
Test: ./gradlew
Change-Id: I8322d3c5993349b52bb4165d7061b55bbf2fe37b