Add tests to show stub constructor selection is order dependent
The selection of the best constructor to call from subclasses using
`super(...)` is done using a partial ordering function which means that
it can compare two different constructors as being equally good in
which case it will pick the first one. That means that the order of
constructors in the input list will affect the selected constructor
which will affect the generated stubs.

That is a problem for two reason:
1. Changes to the order of constructors in the sources can cause
   churn in the generated stubs which in turn causes more work in the
   build.
2. Snapshots can store constructors in a different order to that found
   in the source and so generate different stubs given the same input.

This change adds a new test class to test that the selection order is
not affected by the input order. It tests the same constructors in two
different orders to make sure the result is the same. They assume that
constructors will be selected based on the size of the parameter types
and then the `CallableItem.comparator`.

At the moment some of the tests are broken. They will be fixed in a
follow-up change that updates the selection criteria to match.

Bug: 368021325
Test: ./gradlew
Change-Id: Ia7960666a890bb02856badc099a35e7730bb093f