Track value class primary constructors in K2
With recent K2 updates, finding the primary constructor for a value class now requires going through the KtClass. This constructor can't be used from Java, but since it can be used from Kotlin it makes sense to track to maintain Kotlin source compatibility.

Because there now might be a constructor on the ClassItem when there was not one on the PsiClass, hasImplicitDefaultConstructor needed to be changed to check the constructor count on the ClassItem so it would not add an incorrect default constructor to a value class.

Bug: 369846185
Test: CommonValueClassTest, still passes with aosp/3307337 (updating to latest K2)
Change-Id: I48fdc72a7051b218e9f0268113de112357010c41